// src/Model/iTechRestApi/AuthenticateResponse.ts
//*************************DO NOT MODIFY**************************
//
//THESE FILES ARE AUTOGENERATED WITH A TYPESCRIPT CODE GENERATION TOOL AND ANY MODIFICATIONS MADE HERE WILL BE LOST.
//PLEASE VISIT http://frhagn.github.io/Typewriter/ TO LEARN MORE ABOUT THIS VISUAL STUDIO EXTENSION
//
//*************************DO NOT MODIFY**************************

import { ITechWebAuthenticatedUser } from "./ITechWebAuthenticatedUser";



export class AuthenticateResponse {
    
    public id: string;
    public forename: string;
    public surname: string;
    public username: string;
    public token: string;
    public idleTimeout: number | null;
    public dualAuthenticationEmail: boolean | null;
    public dualAuthenticationSms: boolean | null;
    public dualAuthenticated: boolean | null;
    public dateChallenged: Date | null;
    public authenticatedUser: ITechWebAuthenticatedUser;
    public userType: UserType;
    public landingPage: string;
}

export enum UserType {
    external = 1,
    internal = 2
}